pyproject.toml
[project]
name = "book-lending-system"
version = "0.1.0"
description = "REST API for managing books and loans"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.3.0",
]
